+++ .gitignore
@@ -1,4 +1,4 @@
 *.pyc
 *.cache
-MarkdownEditing.taskpaper
+MarkdownWiki.taskpaper
 /_warehouse
