+++ Markdown (Standard).sublime-settings
@@ -1,12 +1,12 @@
 {
-	"color_scheme": "Packages/MarkdownEditing/MarkdownEditor.tmTheme",
-	// "color_scheme": "Packages/MarkdownEditing/MarkdownEditor-Dark.tmTheme",
-	// "color_scheme": "Packages/MarkdownEditing/MarkdownEditor-Yellow.tmTheme",
+	"color_scheme": "Packages/MarkdownWiki/MarkdownEditor.tmTheme",
+	// "color_scheme": "Packages/MarkdownWiki/MarkdownEditor-Dark.tmTheme",
+	// "color_scheme": "Packages/MarkdownWiki/MarkdownEditor-Yellow.tmTheme",
 
 	// This is a quick and dirty focus theme. In order to make it work, you've to
 	// set `"highlight_line": true,` in this settings file. It is likely that there will
 	// be more mature focus improvements in the future (maybe similar to iA Writer).
-	// "color_scheme": "Packages/MarkdownEditing/MarkdownEditor-Focus.tmTheme",
+	// "color_scheme": "Packages/MarkdownWiki/MarkdownEditor-Focus.tmTheme",
 
 	"tab_size": 4,
 	"translate_tabs_to_spaces": true,
