+++ MarkdownEditor-Yellow.tmTheme
@@ -3,7 +3,7 @@
 <plist version="1.0">
 <dict>
     <key>name</key>
-    <string>MarkdownEditing</string>
+    <string>MarkdownWiki</string>
     <key>settings</key>
     <array>
         <dict>
@@ -688,10 +688,10 @@
             <string>meta.separator.markdown</string>
             <key>settings</key>
             <dict>
-                <key>background</key>
-                <string>#c6b99122</string>
                 <key>foreground</key>
-                <string>#b7a884</string>
+                <string>#a99777</string>
+                <key>background</key>
+                <string>#e2daab</string>
             </dict>
         </dict>
         <dict>
