+++ MarkdownEditor.tmTheme
@@ -3,7 +3,7 @@
 <plist version="1.0">
 <dict>
     <key>name</key>
-    <string>MarkdownEditing</string>
+    <string>MarkdownWiki</string>
     <key>settings</key>
     <array>
         <dict>
@@ -689,9 +689,7 @@
             <key>settings</key>
             <dict>
                 <key>background</key>
-                <string>#BBBBBB22</string>
-                <key>foreground</key>
-                <string>#AAAAAA</string>
+                <string>#BBBBBB</string>
             </dict>
         </dict>
         <dict>
