+++ lint.py
@@ -2,7 +2,7 @@ import sublime
 import sublime_plugin
 import re
 try:
-    from MarkdownEditing.mdeutils import *
+    from MarkdownWiki.mdeutils import *
 except ImportError:
     from mdeutils import *
