+++ tests/Markdown.md
@@ -81,6 +81,8 @@ __ __ __
 
 ----------------
 
+----
+
 
 Block formatting
 ================
